feat: added sidebar
